pyproject.toml
[project]
name = "flights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
    "serpapi>=0.1.5",
]
